imageClassifications/ex_AzureFunctionsPython/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
tensorflow==2.8.2
numpy==1.23.3
Pillow==9.2.0